refactor: Added cross platform abstraction
